update route for fetching music

+++ src/main/java/com/example/MusicPlayer/repo/ViewsRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface ViewsRepo extends JpaRepository<Views,Long> {
     @Query(value = "Select count(*) from views where views.user_id = ?1 AND views.song_id = ?2",nativeQuery = true)
-    int viewExistByUserIdAndSongId(int userId, int songId);
+    int viewExistByUserIdAndSongId(long userId, int songId);
 
 }
